+++ operator/src/lib.rs
@@ -3,6 +3,7 @@
 
 pub mod cli;
 pub mod helpers;
+pub mod programs;
 pub mod prover;
 pub mod rpc;
 pub mod runners;
